Fixed a trivially overlength line.
Sorry, but I want to know what happens if I do down stream edits on top
of yours.  Unlike code, where comments are pushed back to the
originators, the convention with text is to re-edit.

The easy fix would be for you to discard and refork, but there might be
other solutions that generalze better.